kill warnings

+++ dill_pkg.c
@@ -29,7 +29,7 @@ unpack_package(char *package, call_t *t, char **code_p)
     t->call_locs = malloc(sizeof(t->call_locs[0]) * pkg->symbol_count);
     memset(t->call_locs, 0, sizeof(t->call_locs[0]) * pkg->symbol_count);
     for (count = 0; count<pkg->symbol_count; count++) {
-	int call_len;
+	size_t call_len;
 	t->call_locs[count].loc = *((int*)p);
 	t->call_locs[count].xfer_name = (p + sizeof(int));
 	call_len = sizeof(int) + strlen(t->call_locs[count].xfer_name) + 1;
